+++ cmd/attesterinclusion.go
@@ -49,7 +49,7 @@ func init() {
 	attesterFlags(attesterInclusionCmd)
 	attesterInclusionCmd.Flags().Int64("epoch", -1, "the last complete epoch")
 	attesterInclusionCmd.Flags().String("pubkey", "", "the public key of the attester")
-	attesterInclusionCmd.Flags().Int64("index", -1, "the index of the attester")
+	attesterInclusionCmd.Flags().String("index", "", "the index of the attester")
 }
 
 func attesterInclusionBindings() {