+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-dev:
+1.23.0:
   - do not fetch sync committee information for epoch summaries prior to Altair
   - ensure that "attester inclusion" without validator returns appropriate error
   - provide more information in "epoch summary" with verbose flag
+  - add "chain eth1votes"
 
 1.22.0:
   - add "ropsten" to the list of supported networks